[arcane+arccore] Move classes relative to 'MDSpan' from 'arcane_utils' to 'arccore_base'.
