+++ arcane/src/arcane/utils/srcs.cmake
@@ -364,7 +364,6 @@ endif()
 
 if (ARCANE_HAS_ACCELERATOR_API)
   list(APPEND ARCANE_SOURCES
-    MDSpan.cc
     NumArray.cc
     DualUniqueArray.cc
     NumArrayUtils.cc
